Consolidate the getImages and getWorks methods
