+++ catalogue/webapp/services/catalogue/images.ts
@@ -11,15 +11,11 @@ import {
   notFound,
   looksLikeCanonicalId,
   catalogueFetch,
+  catalogueQuery,
+  QueryProps,
 } from '.';
 import { Toggles } from '@weco/toggles';
 import { propsToQuery } from '@weco/common/utils/routes';
-
-type GetImagesProps = {
-  params: CatalogueImagesApiProps;
-  pageSize?: number;
-  toggles: Toggles;
-};
 
 type ImageInclude =
   | 'visuallySimilar'
@@ -34,34 +30,10 @@ type GetImageProps = {
   include?: ImageInclude[];
 };
 
-export async function getImages({
-  params,
-  toggles,
-  pageSize = 25,
-}: GetImagesProps): Promise<CatalogueResultsList<Image> | CatalogueApiError> {
-  const apiOptions = globalApiOptions(toggles);
-
-  const extendedParams = {
-    ...params,
-    pageSize,
-  };
-
-  const searchParams = new URLSearchParams(
-    propsToQuery(extendedParams)
-  ).toString();
-
-  const url = `${
-    rootUris[apiOptions.env]
-  }/v2/images?${searchParams.toString()}`;
-
-  try {
-    const res = await catalogueFetch(url);
-    const json = await res.json();
-
-    return json;
-  } catch (error) {
-    return catalogueApiError();
-  }
+export async function getImages(
+  props: QueryProps<CatalogueImagesApiProps>
+): Promise<CatalogueResultsList<Image> | CatalogueApiError> {
+  return catalogueQuery('images', props);
 }
 
 export async function getImage({
